+++ src/Framework/FieldsAPI/Checkbox.php
@@ -10,6 +10,7 @@ class Checkbox extends Field {
 	use Concerns\HasEmailTag;
 	use Concerns\HasHelpText;
 	use Concerns\HasLabel;
+	use Concerns\HasOptions;
 	use Concerns\HasPlaceholder;
 	use Concerns\ShowInReceipt;
 	use Concerns\StoreAsMeta;